--- tavern_pathfinder/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "tavern_pathfinder"
-version = "0.1.0"
-authors = [
-    "Michael Bryant <shadow53@shadow53.com>",
-    "Benjamin Croom <tauntalus@outlook.com>",
-]
-description = "A library representing all data necessary for representing a complete Pathfinder 1E character"
-documentation = "https://docs.rs/tavern_pathfinder"
-readme = "README.md"
-#homepage = ""
-repository = "https://github.com/d20-tavern/server"
-license-file = "LICENSE"
-keywords = ["pathfinder", "tabletop", "roleplaying", "d20", "characters"]
-#categories = 
-edition = "2018"
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-tavern = [ "tavern_db", "diesel" ]
-
-[dependencies]
-futures = "0.3"
-serde = { version = "1.0.106", features = ["derive"] }
-diesel = { version = "1.4", features = ["postgres", "extras"], optional = true }
-diesel-derive-enum = { version = "1.0", features = ["postgres"] }
-tavern_db = {version = "0.1", path = "../tavern_db", optional = true }
-tavern_derive = {version = "0.1", path = "../tavern_derive"}
-tokio = { version = "0.2", features = [ "blocking", "macros", "stream", "sync" ] }
-uuid = {version = "0.8", features = ["serde"]}
