+++ tavern_server/Cargo.toml
@@ -24,21 +24,22 @@ maintenance = { status = "actively-developed" }
 db-test = [] # Used to conditionally enable db tests
 
 [dependencies]
-async-trait = "0.1"
 base64 = "0.12"
 bytes = "0.5.4"
+diesel = { version = "1.4.4", features = ["postgres", "extras", "r2d2", "uuidv07"] }
+diesel-derive-enum = { version = "1", features = ["postgres"] }
+diesel_migrations = { version ="1.4", features = ["postgres"] }
 futures = "0.3"
 http = "0.2"
-lazy_static = "1.4.0"
+lazy_static = "1.4"
 nebula_form = { version = "0.1", features = ["server-warp"] }
 nebula_status = { version = "0.1", features = ["server-warp"] }
+r2d2 = "0.8"
 rand = "0.7.3"
 rust-argon2 = "0.8"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-sqlx = { version = "0.3", default-features = false, features = [ "chrono", "postgres", "uuid", "runtime-tokio" ] }
 structopt = "0.3"
-tavern_db = { version = "0.1", path = "../tavern_db" }
 tavern_derive = { version = "0.1", path = "../tavern_derive" }
 tokio = { version = "0.2", features = ["full"] }
 uuid = { version = "0.8", features = ["v4", "serde"] }
