+++ tavern_server/src/lib.rs
@@ -1,3 +1,7 @@
+// This macro_use is necessary until diesel 2.0
+#[macro_use]
+extern crate diesel;
+
 use warp::filters::BoxedFilter;
 use warp::{Filter, Reply};
 
@@ -6,7 +10,8 @@ pub mod config;
 pub mod db;
 pub mod forms;
 pub mod status;
-pub(crate) schema;
+pub mod pathfinder;
+mod schema;
 
 /// Generate a warp Filter containing the full server and return it.
 pub fn setup_server() -> BoxedFilter<(impl Reply,)> {
