+++ tavern_server/src/pathfinder/mod.rs
@@ -1,17 +1,3 @@
-use async_trait::async_trait;
-use tavern_pathfinder::summary::{Summary, Summarize};
-use uuid::Uuid;
-use warp::reject::Rejection;
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[test]
-	fn it_works() {
-	}
-}
-
 pub mod character;
 pub mod class;
 pub mod effects;
@@ -19,3 +5,188 @@ pub mod feat;
 pub mod item;
 pub mod religion;
 pub mod spell;
+pub mod summary;
+
+use serde::{Deserialize, Serialize};
+use std::cmp::{Eq, Ord, PartialEq, PartialOrd};
+use std::collections::BTreeMap;
+use std::hash::Hash;
+use tavern_derive::Display;
+use diesel_derive_enum::DbEnum;
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display)]
+pub enum Gender {
+    Male,
+    Female,
+    Other,
+}
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display)]
+pub enum Size {
+    Fine,
+    Diminutive,
+    Tiny,
+    Small,
+    Medium,
+    Large,
+    Huge,
+    Gargantuan,
+    Colossal,
+}
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display)]
+pub enum Alignment {
+    LawfulGood,
+    LawfulNeutral,
+    LawfulEvil,
+    NeutralGood,
+    TrueNeutral,
+    NeutralEvil,
+    ChaoticGood,
+    ChaoticNeutral,
+    ChaoticEvil,
+}
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display, PartialOrd, Ord, Hash, PartialEq, Eq)]
+pub enum Attribute {
+    Strength,
+    Dexterity,
+    Constitution,
+    Intelligence,
+    Wisdom,
+    Charisma,
+}
+
+pub type Attributes = BTreeMap<Attribute, i8>;
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display, PartialOrd, Ord, Hash, PartialEq, Eq)]
+pub enum Skill {
+    Acrobatics,
+    Appraise,
+    Bluff,
+    Climb,
+    Craft,
+    Diplomacy,
+    DisableDevice,
+    Disguise,
+    EscapeArtist,
+    Fly,
+    HandleAnimal,
+    Heal,
+    Intimidate,
+    KnowledgeArcana,
+    KnowledgeDungeoneering,
+    KnowledgeEngineering,
+    KnowledgeGeography,
+    KnowledgeHistory,
+    KnowledgeLocal,
+    KnowledgeNobility,
+    KnowledgePlanes,
+    KnowledgeReligion,
+    Linguistics,
+    Perception,
+    Perform,
+    Profession,
+    Ride,
+    SenseMotive,
+    SleightOfHand,
+    Spellcraft,
+    Stealth,
+    Survival,
+    Swim,
+    UseMagicDevice,
+}
+
+pub type Skills = BTreeMap<Skill, i16>;
+
+#[derive(DbEnum, Debug)]
+// TODO: Revise code to allow for storing data (i16) with the enum instead of separate types
+#[derive(Serialize, Deserialize, Display, PartialOrd, Ord, Hash, PartialEq, Eq)]
+pub enum SaveThrow {
+    Fortitude,
+    Reflex,
+    Will,
+}
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display, PartialOrd, Ord, Hash, PartialEq, Eq)]
+pub enum CharacterStat {
+    Name,
+    Race,
+    Size,
+    Height,
+    Weight,
+    Age,
+    Gender,
+    Alignment,
+    Deity,
+    Languages,
+    Appearance,
+}
+
+pub type CharacterStats = BTreeMap<CharacterStat, i16>;
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display, PartialOrd, Ord, Hash, PartialEq, Eq)]
+pub enum CombatStat {
+    MeleeAttackBonus,
+    RangedAttackBonus,
+    CMB,
+    CMD,
+    ArmorClass,
+    TouchAC,
+    FlatFootedAC,
+    InitiativeBonus,
+    DamageReduction,
+    SpellResistance,
+    Speed,
+    Fortitude,
+    Reflex,
+    Will,
+}
+
+pub type CombatStats = BTreeMap<CombatStat, i16>;
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display)]
+pub enum EquipmentSlot {
+    Armor,
+    Belts,
+    Body,
+    Chest,
+    Eyes,
+    Feet,
+    Hands,
+    Head,
+    Headband,
+    Neck,
+    RingLeft,
+    RingRight,
+    Shield,
+    Shoulders,
+    Wrist,
+}
+
+#[derive(DbEnum, Debug)]
+#[derive(Serialize, Deserialize, Display)]
+pub enum DamageType {
+    Bludgeoning,
+    Slashing,
+    Piercing,
+    Energy,
+    Acid,
+    Fire,
+    Electricity,
+    Cold,
+    Sonic,
+    Positive,
+    Negative,
+    Nonlethal,
+}
+
+pub type Links = BTreeMap<String, String>;
