+++ tavern_server/src/pathfinder/spell.rs
@@ -1,32 +1,23 @@
-use crate::Links;
 use serde::{Deserialize, Serialize};
 use std::cmp::{Eq, Ord, PartialEq, PartialOrd};
-use std::collections::{BTreeSet, BTreeMap};
+use std::collections::BTreeSet;
+use std::convert::TryFrom;
 use uuid::Uuid;
 
-use crate::SaveThrow;
+use super::{Links, SaveThrow};
+use super::effects::{Effect, DBEffect};
+use super::item::{DBItem, Item};
+use super::summary::{Summarize, Summary};
+use crate::schema::{spells, spelleffects, spellcomponents};
 
-#[cfg( feature = "tavern")]
-use crate::effects::DBEffect;
-use crate::effects::Effect;
-#[cfg( feature = "tavern")]
-use crate::item::DBItem;
-use crate::item::Item;
-use crate::summary::{Summarize, Summary};
-#[cfg( feature = "tavern")]
+use crate::db::{GetAll, GetById, Insert, Update, Delete, Error, Connection};
 use diesel_derive_enum::DbEnum;
-#[cfg( feature = "tavern")]
-use tavern_db::{GetAll, GetById, Insert, Update, Delete, Error};
-#[cfg( feature = "tavern")]
-use diesel::PgConnection;
-#[cfg( feature = "tavern")]
-use tavern_db::schema::{spells, spelleffects, spellcomponents};
-#[cfg( feature = "tavern")]
-use std::convert::TryFrom;
+use diesel::prelude::*;
+use crate::db;
 
 #[derive(Serialize, Deserialize, Summarize)]
 pub struct Spell {
-    links: Option<Links>,
+    links: Links,
     id: Uuid,
     components: BTreeSet<SpellComponent>,
     name: String,
@@ -42,11 +33,10 @@ pub struct Spell {
     description: String,
 }
 
-#[cfg(feature = "tavern")]
-impl TryFrom<(&DBSpell, &PgConnection)> for Spell {
-    type Error = tavern_db::Error;
+impl TryFrom<(DBSpell, &Connection)> for Spell {
+    type Error = db::Error;
 
-    fn try_from((db_spell, conn): (DBSpell, &PgConnection)) -> Result<Self, Self::Error> {
+    fn try_from((db_spell, conn): (DBSpell, &Connection)) -> Result<Self, Self::Error> {
         let spell_effects = DBSpellEffect::belonging_to(&db_spell)
             .load::<DBSpellEffect>(conn)
             .map_err(Error::RunQuery)?;
@@ -55,16 +45,16 @@ impl TryFrom<(&DBSpell, &PgConnection)> for Spell {
             .collect()
             .map_err(Error::RunQuery)?;
 
-        let spell_effects = DBSpellComponent::belonging_to(&db_spell)
+        let spell_components = DBSpellComponent::belonging_to(&db_spell)
             .load::<DBSpellComponent>(conn)
             .map_err(Error::RunQuery)?;
-        let effects = spell_effects.into_iter()
-            .map(|spell_effect| SpellComponent::db_get_by_id(spell_effect.effect_id, conn))
+        let components = spell_components.into_iter()
+            .map(|spell_component| SpellComponent::db_get_by_id(spell_component.item_id, conn))
             .collect()
             .map_err(Error::RunQuery)?;
         
         let spell = Spell {
-            links: None,
+            links: Default::default(),
             id: db_spell.id,
             components,
             name: db_spell.name,
@@ -84,21 +74,17 @@ impl TryFrom<(&DBSpell, &PgConnection)> for Spell {
     }
 }
 
-#[cfg_attr(feature = "tavern", derive(GetAll, GetById, Delete, Insert, Update))]
-
-#[cfg(feature = "tavern")]
 impl GetAll for Spell {
-    fn db_get_all(conn: &PgConnection) -> Result<Vec<Self>, Error> {
+    fn db_get_all(conn: &Connection) -> Result<Vec<Self>, Error> {
         let db_spells = DBSpell::db_get_all(conn)?;
-        db.spells.into_iter()
-            .map(|spell| Spell::try_from((spell, conn)))
+        db_spells.into_iter()
+            .map(|db_spell| Spell::try_from((db_spell, conn)))
             .collect()
     }
 }
 
-#[cfg(feature = "tavern")]
 impl GetById for Spell {
-    fn db_get_by_id(id: &Uuid, conn: &PgConnection) -> Result<Self, Error> {
+    fn db_get_by_id(id: &Uuid, conn: &Connection) -> Result<Self, Error> {
         let db_spell = DBSpell::db_get_by_id(id, conn)?;
         Spell::try_from((db_spell, conn))
     }
@@ -111,10 +97,9 @@ impl Insert for Spell {
     }
 }*/
 
-#[cfg(feature = "tavern")]
-#[cfg_attr(feature = "tavern", derive(AsChangeSet, Associations, Identifiable, Insertable, Queryable))]
-#[cfg_attr(feature = "tavern", derive(GetAll, GetById, Delete, Insert, Update))]
-#[cfg_attr(feature = "tavern", table_name = "Spells")]
+#[derive(AsChangeset, Associations, Identifiable, Insertable, Queryable)]
+#[derive(GetAll, GetById, Delete, Insert, Update)]
+#[table_name = "spells"]
 pub struct DBSpell {
     id: Uuid,
     name: String,
@@ -129,13 +114,12 @@ pub struct DBSpell {
     description: String,
 }
 
-#[cfg(feature = "tavern")]
-#[cfg_attr(feature = "tavern", derive(AsChangeSet, Associations, Identifiable, Insertable, Queryable))]
-#[cfg_attr(feature = "tavern", derive(Delete, Insert))]
-#[cfg_attr(feature = "tavern", tavern(id_field = "spell_id"))]
-#[cfg_attr(feature = "tavern", table_name = "SpellEffects")]
-#[cfg_attr(feature = "tavern", belongs_to(DBSpell, foreign_key = "spell_id"))]
-#[cfg_attr(feature = "tavern", belongs_to(DBEffect, foreign_key = "effect_id"))]
+#[derive(Associations, Identifiable, Insertable, Queryable)]
+#[derive(GetById, Delete, Insert)]
+#[tavern(id_field = "spell_id")]
+#[table_name = "spelleffects"]
+#[primary_key(spell_id, effect_id)]
+#[belongs_to(DBSpell, foreign_key = "spell_id")]
 pub struct DBSpellEffect {
     spell_id: Uuid,
     effect_id: Uuid,
@@ -143,6 +127,7 @@ pub struct DBSpellEffect {
 
 #[derive(Serialize, Deserialize, PartialOrd, Ord, PartialEq, Eq)]
 pub struct SpellComponent {
+    id: Uuid,
     item: Option<(Summary<Item>, i16)>,
     description: String,
     amount: i16,
@@ -150,12 +135,10 @@ pub struct SpellComponent {
     component_type: ComponentType,
 }
 
-#[cfg(feature = "tavern")]
-#[cfg_attr(feature = "tavern", derive(AsChangeSet, Associations, Identifiable, Insertable, Queryable))]
-#[cfg_attr(feature = "tavern", derive(GetAll, GetById, Delete, Insert, Update))]
-#[cfg_attr(feature = "tavern", table_name = "SpellComponents")]
-#[cfg_attr(feature = "tavern", belongs_to(DBSpell, foreign_key = "spell_id"))]
-#[cfg_attr(feature = "tavern", belongs_to(DBItem, foreign_key = "item_id"))]
+#[derive(AsChangeset, Associations, Identifiable, Insertable, Queryable)]
+#[derive(GetAll, GetById, Delete, Insert, Update)]
+#[table_name = "spellcomponents"]
+#[belongs_to(DBSpell, foreign_key = "spell_id")]
 pub struct DBSpellComponent {
     id: Uuid,
     spell_id: Uuid,
@@ -165,14 +148,14 @@ pub struct DBSpellComponent {
 }
 
 #[derive(Serialize, Deserialize)]
-#[cfg_attr(feature = "tavern", derive(DbEnum))]
+#[derive(DbEnum, Debug)]
 pub enum CasterType {
     Spontaneous,
     Prepared,
 }
 
 #[derive(Serialize, Deserialize, PartialOrd, Ord, PartialEq, Eq)]
-#[cfg_attr(feature = "tavern", derive(DbEnum))]
+#[derive(DbEnum, Debug)]
 pub enum ComponentType {
     Somatic,
     Material,
@@ -180,7 +163,7 @@ pub enum ComponentType {
 }
 
 #[derive(Serialize, Deserialize)]
-#[cfg_attr(feature = "tavern", derive(DbEnum))]
+#[derive(DbEnum, Debug)]
 pub enum MagicSchool {
     Abjuration,
     Conjuration,
@@ -193,7 +176,7 @@ pub enum MagicSchool {
 }
 
 #[derive(Serialize, Deserialize)]
-#[cfg_attr(feature = "tavern", derive(DbEnum))]
+#[derive(DbEnum, Debug)]
 pub enum SpellRange {
     Personal,
     Touch,
